refactor: common 계층 정리
