+++ src/main/java/com/flab/just_10_minutes/common/convertor/slack/model/SlackMessage.java
@@ -1,4 +1,4 @@
-package com.flab.just_10_minutes.common.alarm.slack;
+package com.flab.just_10_minutes.common.convertor.slack.model;
 
 import com.flab.just_10_minutes.payment.domain.PaymentResultStatus;
 import lombok.Builder;
