+++ src/main/java/com/flab/just_10_minutes/payment/service/PaymentWebhookService.java
@@ -5,7 +5,7 @@ import com.flab.just_10_minutes.payment.dto.PortOnePaidWebhookRequest;
 import com.flab.just_10_minutes.payment.infrastructure.Iamport.IamportApiClient;
 import com.flab.just_10_minutes.payment.infrastructure.entity.PaymentResultEntity;
 import com.flab.just_10_minutes.payment.infrastructure.repository.PaymentResultDao;
-import com.flab.just_10_minutes.common.alarm.slack.SlackClient;
+import com.flab.just_10_minutes.common.convertor.slack.model.SlackClient;
 import com.flab.just_10_minutes.common.exception.iamport.IamportException;
 import com.flab.just_10_minutes.common.exception.webhook.WebhookException;
 import lombok.RequiredArgsConstructor;
@@ -13,8 +13,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import static com.flab.just_10_minutes.payment.domain.PaymentResultStatus.PAID;
-import static com.flab.just_10_minutes.common.alarm.slack.SlackMessage.createDiffersStatus;
-import static com.flab.just_10_minutes.common.alarm.slack.SlackMessage.createImpUidMissingInternal;
+import static com.flab.just_10_minutes.common.convertor.slack.model.SlackMessage.createDiffersStatus;
+import static com.flab.just_10_minutes.common.convertor.slack.model.SlackMessage.createImpUidMissingInternal;
 import static com.flab.just_10_minutes.common.exception.webhook.WebHookMessage.*;
 
 @Slf4j
